add chessboard and move run classes to policy package
