+++ src/main/java/org/matsim/run/policies/OpenBerlinDrtScenario.java
@@ -18,7 +18,7 @@
  *                                                                         *
  * *********************************************************************** */
 
-package org.matsim.run;
+package org.matsim.run.policies;
 
 import ch.sbb.matsim.config.SwissRailRaptorConfigGroup;
 import ch.sbb.matsim.routing.pt.raptor.RaptorIntermodalAccessEgress;
@@ -66,6 +66,7 @@ import org.matsim.legacy.run.drt.OpenBerlinIntermodalPtDrtRouterAnalysisModeIden
 import org.matsim.legacy.run.drt.OpenBerlinIntermodalPtDrtRouterModeIdentifier;
 import org.matsim.pt.transitSchedule.api.TransitSchedule;
 import org.matsim.pt.transitSchedule.api.TransitStopFacility;
+import org.matsim.run.OpenBerlinScenario;
 import picocli.CommandLine;
 
 import java.util.*;
